fix(magnetometer_broadcaster): Refactor to avoid pluginlib in tests
