+++ magnetometer_broadcaster/src/magnetometer_broadcaster.cpp
@@ -17,129 +17,107 @@
  * Authors: Christian Rauch, Subhas Das, Denis Stogl, Victor Lopez
  */
 
-#include <memory>
-#include <vector>
-
-#include <controller_interface/controller_interface.hpp>
+#include <magnetometer_broadcaster/magnetometer_broadcaster.hpp>
 #include <magnetometer_broadcaster/magnetometer_broadcaster_parameters.hpp>
 #include <pluginlib/class_list_macros.hpp>
-#include <rclcpp_lifecycle/state.hpp>
-#include <realtime_tools/realtime_publisher.hpp>
-#include <semantic_components/magnetic_field_sensor.hpp>
-#include <sensor_msgs/msg/magnetic_field.hpp>
 
 namespace magnetometer_broadcaster
 {
-
-class MagnetometerBroadcaster : public controller_interface::ControllerInterface
+controller_interface::InterfaceConfiguration
+MagnetometerBroadcaster::command_interface_configuration() const
 {
-public:
-  controller_interface::InterfaceConfiguration command_interface_configuration() const override
+  return {};
+}
+
+controller_interface::InterfaceConfiguration
+MagnetometerBroadcaster::state_interface_configuration() const
+{
+  return {
+    .type = controller_interface::interface_configuration_type::INDIVIDUAL,
+    .names = magnetometer_->get_state_interface_names(),
+  };
+}
+
+controller_interface::CallbackReturn MagnetometerBroadcaster::on_init()
+{
+  try
   {
-    return {};
+    param_listener_ = std::make_shared<ParamListener>(get_node());
+  }
+  catch (const std::exception & e)
+  {
+    RCLCPP_ERROR_STREAM(
+      get_node()->get_logger(), "Exception thrown during init stage with message: " << e.what());
+    return CallbackReturn::ERROR;
   }
 
-  controller_interface::InterfaceConfiguration state_interface_configuration() const override
+  return CallbackReturn::SUCCESS;
+}
+
+controller_interface::CallbackReturn MagnetometerBroadcaster::on_configure(
+  const rclcpp_lifecycle::State & /*previous_state*/)
+{
+  try
   {
-    return {
-      .type = controller_interface::interface_configuration_type::INDIVIDUAL,
-      .names = magnetometer_->get_state_interface_names(),
-    };
+    params_ = param_listener_->get_params();
+  }
+  catch (const std::exception & e)
+  {
+    RCLCPP_ERROR_STREAM(
+      get_node()->get_logger(), "Exception thrown during config stage with message: " << e.what());
+    return CallbackReturn::ERROR;
   }
 
-  controller_interface::CallbackReturn on_init() override
+  magnetometer_ = std::make_unique<semantic_components::MagneticFieldSensor>(params_.sensor_name);
+  try
   {
-    try
-    {
-      param_listener_ = std::make_shared<ParamListener>(get_node());
-    }
-    catch (const std::exception & e)
-    {
-      RCLCPP_ERROR_STREAM(
-        get_node()->get_logger(), "Exception thrown during init stage with message: " << e.what());
-      return CallbackReturn::ERROR;
-    }
-
-    return CallbackReturn::SUCCESS;
+    sensor_state_publisher_ = get_node()->create_publisher<sensor_msgs::msg::MagneticField>(
+      "~/magnetic_field", rclcpp::SystemDefaultsQoS());
+    realtime_publisher_ = std::make_unique<StatePublisher>(sensor_state_publisher_);
+  }
+  catch (const std::exception & e)
+  {
+    RCLCPP_ERROR_STREAM(
+      get_node()->get_logger(),
+      "Exception thrown during publisher creation at configure stage with message: " << e.what());
+    return CallbackReturn::ERROR;
   }
 
-  controller_interface::CallbackReturn on_configure(
-    const rclcpp_lifecycle::State & /*previous_state*/) override
+  state_message_.header.frame_id = params_.frame_id;
+  std::copy(
+    params_.static_covariance.begin(), params_.static_covariance.end(),
+    state_message_.magnetic_field_covariance.begin());
+
+  return CallbackReturn::SUCCESS;
+}
+
+controller_interface::CallbackReturn MagnetometerBroadcaster::on_activate(
+  const rclcpp_lifecycle::State & /*previous_state*/)
+{
+  magnetometer_->assign_loaned_state_interfaces(state_interfaces_);
+  return CallbackReturn::SUCCESS;
+}
+
+controller_interface::CallbackReturn MagnetometerBroadcaster::on_deactivate(
+  const rclcpp_lifecycle::State & /*previous_state*/)
+{
+  magnetometer_->release_interfaces();
+  return CallbackReturn::SUCCESS;
+}
+
+controller_interface::return_type MagnetometerBroadcaster::update(
+  const rclcpp::Time & time, const rclcpp::Duration & /*period*/)
+{
+  magnetometer_->get_values_as_message(state_message_);
+
+  if (realtime_publisher_)
   {
-    try
-    {
-      params_ = param_listener_->get_params();
-    }
-    catch (const std::exception & e)
-    {
-      RCLCPP_ERROR_STREAM(
-        get_node()->get_logger(),
-        "Exception thrown during config stage with message: " << e.what());
-      return CallbackReturn::ERROR;
-    }
-
-    magnetometer_ = std::make_unique<semantic_components::MagneticFieldSensor>(params_.sensor_name);
-    try
-    {
-      sensor_state_publisher_ = get_node()->create_publisher<sensor_msgs::msg::MagneticField>(
-        "~/magnetic_field", rclcpp::SystemDefaultsQoS());
-      realtime_publisher_ = std::make_unique<StatePublisher>(sensor_state_publisher_);
-    }
-    catch (const std::exception & e)
-    {
-      RCLCPP_ERROR_STREAM(
-        get_node()->get_logger(),
-        "Exception thrown during publisher creation at configure stage with message: " << e.what());
-      return CallbackReturn::ERROR;
-    }
-
-    state_message_.header.frame_id = params_.frame_id;
-    std::copy(
-      params_.static_covariance.begin(), params_.static_covariance.end(),
-      state_message_.magnetic_field_covariance.begin());
-
-    return CallbackReturn::SUCCESS;
+    state_message_.header.stamp = time;
+    realtime_publisher_->try_publish(state_message_);
   }
 
-  controller_interface::CallbackReturn on_activate(
-    const rclcpp_lifecycle::State & /*previous_state*/) override
-  {
-    magnetometer_->assign_loaned_state_interfaces(state_interfaces_);
-    return CallbackReturn::SUCCESS;
-  }
-
-  controller_interface::CallbackReturn on_deactivate(
-    const rclcpp_lifecycle::State & /*previous_state*/) override
-  {
-    magnetometer_->release_interfaces();
-    return CallbackReturn::SUCCESS;
-  }
-
-  controller_interface::return_type update(
-    const rclcpp::Time & time, const rclcpp::Duration & /*period*/) override
-  {
-    magnetometer_->get_values_as_message(state_message_);
-
-    if (realtime_publisher_)
-    {
-      state_message_.header.stamp = time;
-      realtime_publisher_->try_publish(state_message_);
-    }
-
-    return controller_interface::return_type::OK;
-  }
-
-protected:
-  std::shared_ptr<ParamListener> param_listener_;
-  Params params_;
-
-  std::unique_ptr<semantic_components::MagneticFieldSensor> magnetometer_;
-
-  using StatePublisher = realtime_tools::RealtimePublisher<sensor_msgs::msg::MagneticField>;
-  rclcpp::Publisher<sensor_msgs::msg::MagneticField>::SharedPtr sensor_state_publisher_;
-  std::unique_ptr<StatePublisher> realtime_publisher_;
-  sensor_msgs::msg::MagneticField state_message_;
-};
+  return controller_interface::return_type::OK;
+}
 
 }  // namespace magnetometer_broadcaster
